Fixes reading files larger than 128K from exFAT.

+++ software/filesystem/partition.cc
@@ -51,7 +51,7 @@ FileSystem *Partition :: attach_filesystem(void)
 	return NULL;
 }
     
-DRESULT Partition::read(uint8_t *buffer, uint32_t sector, uint8_t count)
+DRESULT Partition::read(uint8_t *buffer, uint32_t sector, int count)
 {
 	if(!dev)
         return RES_NOTRDY;
@@ -61,7 +61,7 @@ DRESULT Partition::read(uint8_t *buffer, uint32_t sector, uint8_t count)
 }
 
 #if	_READONLY == 0
-DRESULT Partition::write(const uint8_t *buffer, uint32_t sector, uint8_t count)
+DRESULT Partition::write(const uint8_t *buffer, uint32_t sector, int count)
 {
     if(!dev)
         return RES_NOTRDY;
